Speech recognition for several events
transition complete, do_land_start, guided alt changes, Detransition, land descent and land final, takeoff.

Vtol Takeoff currently does not get read out loud due to only announcing when the aircraft is flying and armed.

If we don't have these checks some of these dialogues can be announced if the waypoint switches on the ground.

These dialogues are not garunteed to announce. If a message is lost there will be no announcement.