Fix Video popout and improve PIP status.
Video popup was broken on windows prior to this update.
Removes the ability to pop out the map screen when video is resized in PIP mode.
Ensures the map screen defaults to the main view no matter the state of the PIP or Windowed modes.
Adds click to track to windowed mode.